feat: Add script to archive events with a transformed schema
index.js will grab events that are in the past and archive them using
the schema defined in lib/schema.js. By updating the schema, it's
possible to add or remove keys from the output event.

The schema is used to populate a base object with default values in all
fields if the fields are missing or empty in the live object. Then some
fields are overwritten with custom transforms.